requirements.txt
pytest
flask
python-dotenv
pytz
ics
flake8
